feat: add simple avg() aggregate function for the embedding type (#33)
We can’t fully vectorize this function using np.nanmean, because we
don’t want to ignore NaN values in the embeddings. If an embedding
contains a NaN, the mean should also be NaN—not because that's
particularly useful, but because it clearly signals that the embedding
is invalid.

We'll add a helper function to flag rows with NaN values in the
embedding column so they can be filtered out before performing any
embedding-related computations.